Tweaked the CSS a bit.
the Summary Box and Spacers were displaying on separate rows. I tweaked the CSS until it worked. 
I am still not happy with the solution, but it will work for now. My main concern is that it might not
work properly on mobile browser screens. I won't be able to test that well until I get the initial live
version on the server where I can load it on my phone.